Resolve $HOME in profile.d at runtime rather than during the build
Previously the value for `$HOME` was resolved at build time, meaning its
actual value during the build is hardcoded in the `.profile.d/` script,
rather than just the variable name.

Whilst the value for `$HOME` is currently the same at both build-time
and run-time (both `/app`), the build-time value (only) is due to change
in the future to a path under `/tmp`.

By escaping the dollar sign, the variable `$HOME` is now resolved at
runtime, making the buildpack compatible with this future change.